Billion things
There's a billion things that changed here. This is definitely not good
practice. But I did it. The codebase is still young, though, so this
should not be a problem beyond what my mind makes of it...

Some things that changed:
- setup.py and all that pip packaging stuff
- mark memory test as slow
- read from text streams rather than byte streams
- allow the Reader to read from multiple files in sequence
- greatly simplify the code by not caching the positions of items in a
  node. This changed the way the package works, but for the best, I
  think.
- remove the ability to randomly seek the files. This is in line with
  the previous point, as well as the idea of not using more memory than
  actually needed. Processing of the JSON files is expected to be
  completely sequential!
- implement `in` membership tests for arrays (mimicking what existined
  objects)
- implement lazy equality of arrays and objects
- remove `Node.keys()`

We're missing lazy operations on strings, but I think I'll add them
later, if needed.