Handle quotes at start of buffers
On measuring strings, we read text in chunks, using a buffer with a
constant specific length, and then decide whether an existing quote is a
termination of the string or part of the string by peeking at the
character before the quote (a quote escaped with a backslash is part of
the string); because the quote can be at the beginning of the buffer, we
must hold on to whether the last character of the previous chunk was a
backslash.

The code that determines whether the quote is part of the string used
the condition

```
(quote_pos == 0 and prev_backslash) or buffer[quote_pos - 1] == '\\'
```

but when `quote_pos` is 0, buffer[quote_pos - 1] tests the last
character of the buffer, which might even fall entirely outside of the
string.

This commit corrects this error.